requirements/dev.txt
-r requirements/common.txt

pyflakes==1.5.0
pytest==3.0.7
pytest-cov==2.4.0
pytest-flask==0.10.0
pytest-flakes==1.0.1
pytest-mock==1.5.0

ipdb==0.10.2
